requirements.txt
# ===== Core: Multi-Agent Workflow =====
langgraph>=0.2.0
langchain-core>=0.3.0

# ===== LLM Providers =====
anthropic>=0.39.0
openai>=1.54.0            # 兼容 DeepSeek / Groq / 硅基流动 / 千问
ollama>=0.3.0             # 本地模型

# ===== Data & Plotting =====
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.13.0

# ===== Backend =====
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12  # 文件上传
sse-starlette>=2.1.0      # SSE 流式推送

# ===== Utilities =====
pydantic>=2.0.0
python-dotenv>=1.0.0
